+++ packages/solana/lib/src/exceptions/rpc_timeout_exception.dart
@@ -1,0 +1,12 @@
+import 'dart:async';
+
+class RpcTimeoutException extends TimeoutException {
+  RpcTimeoutException({
+    required String method,
+    required this.body,
+    Duration? timeout,
+  }) : super('RPC call $method timed out.', timeout);
+
+  /// Request body.
+  final dynamic body;
+}